network_vis_postiion_chart
Code in python using networkx
Generates 9 random jobs, assigns them random salaries and skill similarities, and picks one to start as the current position. From there it illustrates the ease of transition from one position to another as well as a slight idea of salary differential. The current position is highlighted to increase readability as a management conversion.